# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorBody:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorDetail'
        type: array
      message:
        type: string
      requestId:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope:
    properties:
      error:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorBody'
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.CameraConfig:
    properties:
      distance:
        type: number
      fov:
        type: number
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.CoreConfig:
    properties:
      color:
        type: string
      emissive:
        type: string
      scale:
        type: number
      shape:
        type: string
      spinSpeed:
        type: number
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.CreateWorldResponse:
    properties:
      personalityDNA:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.PersonalityDNA'
      variant:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.WorldVariant'
      world:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.World'
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.DNAPlanet:
    properties:
      energy:
        type: integer
      key:
        type: string
      meaning:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.EnergySignature:
    properties:
      intensity:
        type: integer
      primary:
        type: string
      secondary:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.HUDConfig:
    properties:
      showLabels:
        type: boolean
      showTraitBars:
        type: boolean
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.Palette:
    properties:
      accent:
        type: string
      background:
        type: string
      gradient:
        items:
          type: string
        type: array
      primary:
        type: string
      secondary:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.ParticleConfig:
    properties:
      color:
        type: string
      desktopCount:
        type: integer
      mobileCount:
        type: integer
      spread:
        type: number
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.PersonalityDNA:
    properties:
      archetype:
        type: string
      energySignature:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.EnergySignature'
      planets:
        items:
          $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.DNAPlanet'
        type: array
      quote:
        type: string
      sceneName:
        type: string
      schemaVersion:
        type: string
      shortNarrative:
        type: string
      traitScores:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.TraitScores'
      visualHints:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.VisualHints'
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.PlanetSceneConfig:
    properties:
      color:
        type: string
      energy:
        type: integer
      key:
        type: string
      meaning:
        type: string
      name:
        type: string
      orbitRadius:
        type: number
      orbitSpeed:
        type: number
      phase:
        type: number
      size:
        type: number
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.PostFXConfig:
    properties:
      bloomIntensity:
        type: number
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.PublicDNA:
    properties:
      planets:
        items:
          $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.DNAPlanet'
        type: array
      traitScores:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.TraitScores'
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.PublicVariant:
    properties:
      config:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.WorldSceneConfig'
      seed:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.PublicWorld:
    properties:
      archetype:
        type: string
      nickname:
        type: string
      quote:
        type: string
      sceneName:
        type: string
      shortNarrative:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.PublicWorldResponse:
    properties:
      publicDNA:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.PublicDNA'
      variant:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.PublicVariant'
      world:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.PublicWorld'
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.PublishResponse:
    properties:
      shareSlug:
        type: string
      shareUrl:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.TraitScores:
    properties:
      creativity:
        type: integer
      curiosity:
        type: integer
      discipline:
        type: integer
      energy:
        type: integer
      focus:
        type: integer
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.VariantResponse:
    properties:
      variant:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.WorldVariant'
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.VisualHints:
    properties:
      coreSymbol:
        type: string
      motionIntent:
        type: string
      paletteIntent:
        type: string
      theme:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.World:
    properties:
      archetype:
        type: string
      createdAt:
        type: string
      id:
        type: string
      nickname:
        type: string
      quote:
        type: string
      role:
        type: string
      sceneName:
        type: string
      selectedVariantId:
        type: string
      shareSlug:
        type: string
      shortNarrative:
        type: string
      updatedAt:
        type: string
      visibility:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.WorldInput:
    properties:
      challenge:
        type: string
      favoriteColors:
        items:
          type: string
        type: array
      goal:
        type: string
      interests:
        items:
          type: string
        type: array
      mood:
        type: string
      nickname:
        type: string
      preferredWorldStyle:
        type: string
      role:
        type: string
      traits:
        items:
          type: string
        type: array
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.WorldResponse:
    properties:
      personalityDNA:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.PersonalityDNA'
      selectedVariant:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.WorldVariant'
      variants:
        items:
          $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.WorldVariant'
        type: array
      world:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.World'
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.WorldSceneConfig:
    properties:
      archetype:
        type: string
      camera:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.CameraConfig'
      core:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.CoreConfig'
      hud:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.HUDConfig'
      palette:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.Palette'
      particles:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.ParticleConfig'
      planets:
        items:
          $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.PlanetSceneConfig'
        type: array
      postFX:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.PostFXConfig'
      quote:
        type: string
      sceneName:
        type: string
      schemaVersion:
        type: string
      theme:
        type: string
    type: object
  github_com_myunivokai_myunivokai_apps_api_internal_models.WorldVariant:
    properties:
      config:
        $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.WorldSceneConfig'
      createdAt:
        type: string
      id:
        type: string
      isSelected:
        type: boolean
      seed:
        type: string
      thumbnailUrl:
        type: string
      variantNo:
        type: integer
      worldId:
        type: string
    type: object
info:
  contact: {}
  description: Backend API for Myunivokai personal universe generation.
  title: Myunivokai API
  version: 0.1.0
paths:
  /healthz:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Health check
      tags:
      - health
  /share/worlds/{shareSlug}:
    get:
      parameters:
      - description: Share slug
        in: path
        name: shareSlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.PublicWorldResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
      summary: Get public universe
      tags:
      - share
  /worlds:
    post:
      consumes:
      - application/json
      parameters:
      - description: World input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.WorldInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.CreateWorldResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
      summary: Create universe
      tags:
      - worlds
  /worlds/{worldId}:
    get:
      parameters:
      - description: World ID
        in: path
        name: worldId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.WorldResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
      summary: Get universe
      tags:
      - worlds
  /worlds/{worldId}/publish:
    post:
      parameters:
      - description: World ID
        in: path
        name: worldId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.PublishResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
      summary: Publish universe
      tags:
      - worlds
  /worlds/{worldId}/variants:
    post:
      parameters:
      - description: World ID
        in: path
        name: worldId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.VariantResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
      summary: Regenerate universe variant
      tags:
      - worlds
  /worlds/{worldId}/variants/{variantId}/select:
    post:
      parameters:
      - description: World ID
        in: path
        name: worldId
        required: true
        type: string
      - description: Variant ID
        in: path
        name: variantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_models.VariantResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_myunivokai_myunivokai_apps_api_internal_httpx.ErrorEnvelope'
      summary: Select universe variant
      tags:
      - worlds
swagger: "2.0"